Consolidate role-check middlewares into single ensure_role.ts file
