+++ front-api/routes/w/[wId]/mcp/views/[viewId]/index.ts
@@ -6,7 +6,7 @@ import {
 import { MCPServerViewResource } from "@app/lib/resources/mcp_server_view_resource";
 import { assertNever } from "@app/types/shared/utils/assert_never";
 import { workspaceApp } from "@front-api/middlewares/ctx";
-import { ensureIsUser } from "@front-api/middlewares/ensure_is_user";
+import { ensureIsUser } from "@front-api/middlewares/ensure_role";
 import type { HandlerResult } from "@front-api/middlewares/utils";
 import { apiError } from "@front-api/middlewares/utils";
 import { validate } from "@front-api/middlewares/validator";
