+++ front-api/routes/w/[wId]/spaces/[spaceId]/webhook_source_views/[webhookSourceViewId]/index.ts
@@ -2,7 +2,7 @@ import { WebhookSourcesViewResource } from "@app/lib/resources/webhook_sources_v
 import { normalizeError } from "@app/types/shared/utils/error_utils";
 import type { SpaceKind } from "@app/types/space";
 import { workspaceApp } from "@front-api/middlewares/ctx";
-import { ensureIsUser } from "@front-api/middlewares/ensure_is_user";
+import { ensureIsUser } from "@front-api/middlewares/ensure_role";
 import type { HandlerResult } from "@front-api/middlewares/utils";
 import { apiError } from "@front-api/middlewares/utils";
 import { withSpace } from "@front-api/middlewares/with_space";
